move gamma components into dedicated directories
