+++ src/components/gamma/split-view/SplitViewHost.types.ts
