+++ src/components/gamma/split-view/SplitViewScreen.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
-import SplitViewScreenNativeComponent from '../../fabric/gamma/SplitViewScreenNativeComponent';
+import SplitViewScreenNativeComponent from '../../../fabric/gamma/SplitViewScreenNativeComponent';
 import { SplitViewScreenProps } from './SplitViewScreen.types';
 
 /**
