+++ src/components/gamma/split-view/SplitViewScreen.types.ts
