+++ src/components/gamma/split-view/SplitViewScreen.web.tsx
