+++ src/components/gamma/stack/StackScreen.types.ts
