+++ src/components/gamma/stack/StackScreen.web.tsx
