+++ src/experimental/index.ts
@@ -8,10 +8,10 @@ export * from './types';
 
 // Components
 
-export { default as ScreenStackHost } from '../components/gamma/ScreenStackHost';
+export { default as ScreenStackHost } from '../components/gamma/stack/ScreenStackHost';
 export {
   default as StackScreen,
   StackScreenLifecycleState,
-} from '../components/gamma/StackScreen';
-export { default as SplitViewHost } from '../components/gamma/SplitViewHost';
-export { default as SplitViewScreen } from '../components/gamma/SplitViewScreen';
+} from '../components/gamma/stack/StackScreen';
+export { default as SplitViewHost } from '../components/gamma/split-view/SplitViewHost';
+export { default as SplitViewScreen } from '../components/gamma/split-view/SplitViewScreen';
