+++ src/experimental/types.ts
@@ -3,7 +3,7 @@
  * BE SUBJECT TO BREAKING CHANGES WITHOUT NOTICE OR LIBRARY MAJOR VERSION CHANGE.
  */
 
-export * from '../components/gamma/SplitViewHost.types';
-export * from '../components/gamma/SplitViewScreen.types';
-export * from '../components/gamma/ScreenStackHost.types';
-export * from '../components/gamma/StackScreen.types';
+export * from '../components/gamma/split-view/SplitViewHost.types';
+export * from '../components/gamma/split-view/SplitViewScreen.types';
+export * from '../components/gamma/stack/ScreenStackHost.types';
+export * from '../components/gamma/stack/StackScreen.types';
